other(core): home dir env override
